Data
The Data here provides a CSV file wich can be downloaded from https://exoplanetarchive.ipac.caltech.edu/cgi-bin/TblView/nph-tblView?app=ExoTbls&config=PS by making click on download table on a CSV format. 
We also add the API connection for those who want to make a query to the server, but we have to say that the objective of this project is to be able to work with no internet connection needed all the time.